// NSDL_FVCI_UAT_RELEASE/13-03-2025/Website Deployment/nsdlfpi-development/src/swagger/model/draftFvciComplianceOfficerDetailsDto.ts
/**
 * NSDL API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface DraftFvciComplianceOfficerDetailsDto { 
    id?: number;
    fvciApplicationId?: string;
    name?: string;
    jobTitle?: string;
    phoneNumber?: string;
    faxNumber?: string;
    status?: number;
    createdAt?: Date;
    updatedAt?: Date;
}